requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
channels>=4.0,<5.0
channels-redis>=4.2,<5.0
daphne>=4.1,<5.0
redis>=5.0,<6.0
djangorestframework-simplejwt>=5.3,<6.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
